# docs/openapi.json
{
    "openapi": "3.0.0",
    "paths": {
...
        "/api/v1/auth/signup": {
            "post": {
                "description": "Create a new user account with email/phone and password.",
                "operationId": "AuthController_signup",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "required": [
                                    "displayName",
                                    "password"
                                ],
                                "properties": {
                                    "displayName": {
                                        "type": "string",
                                        "example": "Rahul Sharma"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "rahul@example.com"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "+919876543210"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "StrongPass123!"
                                    },
                                    "deviceLabel": {
                                        "type": "string",
                                        "example": "Chrome / macOS"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User successfully created & tokens issued"
                    },
                    "400": {
                        "description": "Validation failed or email/phone already in use"
                    }
                },
                "summary": "Register User",
                "tags": [
                    "Auth"
                ]
            }
        },
        "/api/v1/auth/login": {
            "post": {
                "description": "Authenticate user with email or phone and password.",
                "operationId": "AuthController_login",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "required": [
                                    "identifier",
                                    "password"
                                ],
                                "properties": {
                                    "identifier": {
                                        "type": "string",
                                        "example": "rahul@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "StrongPass123!"
                                    },
                                    "deviceLabel": {
                                        "type": "string",
                                        "example": "iPhone 15 Pro App"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login successful, returns access & refresh tokens"
                    },
                    "401": {
                        "description": "Invalid credentials"
                    }
                },
                "summary": "Password Login",
                "tags": [
                    "Auth"
                ]
            }
        },
...
        "/api/v1/auth/refresh": {
            "post": {
                "description": "Issue a new access token using a valid refresh token.",
                "operationId": "AuthController_refresh",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "required": [
                                    "refreshToken"
                                ],
                                "properties": {
                                    "refreshToken": {
                                        "type": "string",
                                        "example": "sample_refresh_token_string"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Tokens rotated successfully"
                    },
                    "401": {
                        "description": "Invalid or revoked refresh token"
                    }
                },
                "summary": "Refresh Access Token",
                "tags": [
                    "Auth"
                ]
            }
        },
...
        "/api/v1/auth/bootstrap": {
            "get": {
                "description": "Get profile info, active permissions, and user status.",
                "operationId": "AuthController_bootstrap",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "User state and permissions payload"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "summary": "Bootstrap Session",
                "tags": [
                    "Auth"
                ]
            }
        },
...
    },
    "info": {
        "title": "College Rental Platform API",
        "description": "Interactive OpenAPI documentation for College Rental Platform backend APIs.",
        "version": "1.0.0",
        "contact": {}
    },
...
    "servers": [],
    "components": {
        "securitySchemes": {
            "bearer": {
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "type": "http",
                "description": "Enter JWT Access Token"
            },
            "admin-key": {
                "type": "apiKey",
                "in": "header",
                "name": "x-admin-key",
                "description": "Admin API Key"
            }
        },
...
    }
}